@@ -4,13 +4,13 @@
   "description": "Static website",
   "private": true,
   "engines": {
-    "yarn": "^1.15",
-    "node": "^10.15 || ^12.6"
+    "yarn": "^1.19",
+    "node": "^10.16 || ^12.6"
   },
   "scripts": {
     "start": "node server/index.js",
     "dev": "webpack-dev-server --host 0.0.0.0 --mode development --color --env.noclear --env.hmr",
-    "build": "webpack --display-error-details --mode development",
+    "build": "cross-env TS_NODE_PROJECT=tsconfig.webpack.json webpack --display-error-details --mode development",
     "build:minify": "yarn build --env.fullminify",
     "build:prod": "webpack --display-error-details --mode production",
     "build:prod:minify": "yarn build:prod --env.fullminify",
@@ -18,54 +18,72 @@
     "heroku-postbuild": "yarn build:prod:minify && cd server && yarn"
   },
   "dependencies": {
-    "@babel/polyfill": "^7.4.4",
+    "@babel/polyfill": "^7.6.0",
     "core-js": "^3",
-    "detect-browser": "^4.5.0",
+    "detect-browser": "^4.7.0",
     "gsap": "^2.1.3",
-    "mobx": "^5.9.4",
-    "regenerator-runtime": "^0.13.2",
+    "mobx": "^5.14.0",
+    "regenerator-runtime": "^0.13.3",
     "reset-css": "^4.0.1"
   },
   "devDependencies": {
-    "@babel/cli": "^7.4.4",
-    "@babel/core": "^7.4.5",
-    "@babel/plugin-proposal-class-properties": "^7.4.4",
-    "@babel/plugin-proposal-decorators": "^7.4.4",
+    "@babel/cli": "^7.6.2",
+    "@babel/core": "^7.6.2",
+    "@babel/plugin-proposal-class-properties": "^7.5.5",
+    "@babel/plugin-proposal-decorators": "^7.6.0",
     "@babel/plugin-syntax-dynamic-import": "^7.2.0",
-    "@babel/preset-env": "^7.4.5",
+    "@babel/preset-env": "^7.6.2",
     "@babel/preset-react": "^7.0.0",
+    "@types/core-js": "^2.5.2",
     "@types/gsap": "^1.20.2",
-    "autoprefixer": "^9.5.1",
-    "babel-eslint": "^10.0.1",
+    "@types/html-webpack-plugin": "^3.2.1",
+    "@types/mini-css-extract-plugin": "^0.8.0",
+    "@types/node": "^12.7.11",
+    "@types/optimize-css-assets-webpack-plugin": "^5.0.1",
+    "@types/ts-nameof": "^4.0.0",
+    "@types/webpack": "^4.39.2",
+    "@types/webpack-dev-server": "^3.1.7",
+    "@typescript-eslint/eslint-plugin": "^2.3.3",
+    "@typescript-eslint/parser": "^2.3.3",
+    "autoprefixer": "^9.6.4",
+    "babel-eslint": "^10.0.3",
     "babel-loader": "^8.0.6",
     "babel-minify-webpack-plugin": "^0.3.1",
-    "clean-webpack-plugin": "^2.0.2",
-    "cross-env": "^5.2.0",
-    "css-loader": "^2.1.1",
-    "eslint": "^5.16.0",
-    "eslint-config-airbnb": "^17.1.0",
-    "eslint-config-prettier": "^4.3.0",
+    "clean-webpack-plugin": "^3.0.0",
+    "cross-env": "^6.0.3",
+    "css-loader": "^3.2.0",
+    "eslint": "^6.5.1",
+    "eslint-config-airbnb": "^18.0.1",
+    "eslint-config-prettier": "^6.4.0",
     "eslint-import-resolver-webpack": "^0.11.1",
-    "eslint-plugin-import": "^2.17.2",
-    "eslint-plugin-jsx-a11y": "^6.2.1",
-    "eslint-plugin-react": "^7.13.0",
-    "file-loader": "^3.0.1",
+    "eslint-plugin-import": "^2.18.2",
+    "eslint-plugin-jsx-a11y": "^6.2.3",
+    "eslint-plugin-react": "^7.16.0",
+    "file-loader": "^4.2.0",
     "html-loader": "^0.5.5",
     "html-webpack-plugin": "^3.2.0",
-    "image-webpack-loader": "^4.6.0",
-    "mini-css-extract-plugin": "^0.6.0",
+    "image-webpack-loader": "^6.0.0",
+    "mini-css-extract-plugin": "^0.8.0",
     "node-sass": "^4.12.0",
-    "optimize-css-assets-webpack-plugin": "^5.0.1",
+    "optimize-css-assets-webpack-plugin": "^5.0.3",
     "postcss-loader": "^3.0.0",
-    "postcss-preset-env": "^6.6.0",
-    "raw-loader": "^1",
+    "postcss-preset-env": "^6.7.0",
+    "raw-loader": "^3",
     "resolve-url-loader": "^3.1.0",
-    "sass-loader": "^7.1.0",
-    "style-loader": "^0.23.1",
+    "sass-loader": "^8.0.0",
+    "style-loader": "^1.0.0",
+    "ts-loader": "^6.2.0",
+    "ts-nameof": "^4.0.3",
+    "ts-node": "^8.4.1",
+    "tsconfig-paths": "^3.9.0",
+    "tslint": "^5.20.0",
+    "tslint-config-airbnb": "^5.11.2",
+    "ttypescript": "^1.5.7",
+    "typescript": "^3.6.3",
     "underscore-template-loader": "^1.0.0",
-    "url-loader": "^1.1.2",
-    "webpack": "^4.32.2",
-    "webpack-cli": "^3.3.2",
-    "webpack-dev-server": "^3.4.1"
+    "url-loader": "^2.2.0",
+    "webpack": "^4.41.0",
+    "webpack-cli": "^3.3.9",
+    "webpack-dev-server": "^3.8.2"
   }
 }
